requirements.txt
django==4.2
ipython
django-extensions
Pillow
requests